Conflicts:
	himan-lib/himan-lib.spec

+++ himan-lib/source/stereographic_grid.cpp
@@ -49,7 +49,7 @@ stereographic_grid::stereographic_grid(const stereographic_grid& other)
 
 size_t stereographic_grid::Size() const
 {
-	if (itsNi == kHPMissingInt || itsNj == kHPMissingInt)
+	if (itsNi == static_cast<size_t> (kHPMissingInt) || itsNj == static_cast<size_t> (kHPMissingInt))
 	{
 		return kHPMissingInt;
 	}
@@ -58,16 +58,27 @@ size_t stereographic_grid::Size() const
 }
 
 void stereographic_grid::Ni(size_t theNi) { itsNi = theNi; }
+
 void stereographic_grid::Nj(size_t theNj) { itsNj = theNj; }
+
 size_t stereographic_grid::Ni() const { return itsNi; }
+
 size_t stereographic_grid::Nj() const { return itsNj; }
+
 void stereographic_grid::Di(double theDi) { itsDi = theDi; }
+
 void stereographic_grid::Dj(double theDj) { itsDj = theDj; }
+
 double stereographic_grid::Di() const { return itsDi; }
+
 double stereographic_grid::Dj() const { return itsDj; }
+
 void stereographic_grid::Orientation(double theOrientation) { itsOrientation = theOrientation; }
+
 double stereographic_grid::Orientation() const { return itsOrientation; }
+
 HPScanningMode stereographic_grid::ScanningMode() const { return itsScanningMode; }
+
 void stereographic_grid::ScanningMode(HPScanningMode theScanningMode)
 {
 	if (theScanningMode != kBottomLeft)
@@ -115,9 +126,13 @@ point stereographic_grid::LatLon(size_t locationIndex) const
 }
 
 void stereographic_grid::BottomLeft(const point& theBottomLeft) { itsBottomLeft = theBottomLeft; }
+
 void stereographic_grid::TopRight(const point& theTopRight) { itsTopRight = theTopRight; }
+
 point stereographic_grid::BottomLeft() const { return itsBottomLeft; }
+
 point stereographic_grid::TopRight() const { return itsTopRight; }
+
 point stereographic_grid::FirstPoint() const
 {
 	switch (itsScanningMode)
@@ -185,6 +200,7 @@ bool stereographic_grid::Swap(HPScanningMode newScanningMode)
 }
 
 stereographic_grid* stereographic_grid::Clone() const { return new stereographic_grid(*this); }
+
 ostream& stereographic_grid::Write(std::ostream& file) const
 {
 	grid::Write(file);
@@ -197,6 +213,7 @@ ostream& stereographic_grid::Write(std::ostream& file) const
 }
 
 bool stereographic_grid::operator!=(const grid& other) const { return !(other == *this); }
+
 bool stereographic_grid::operator==(const grid& other) const
 {
 	const stereographic_grid* g = dynamic_cast<const stereographic_grid*>(&other);
